add permission for member operation

+++ application/views/pages/member_list.tpl
@@ -13,10 +13,12 @@
 						<td><%$value%></td>
 					<%/if%>
 				<%/foreach%>
-				<td>
-					<a href='<%$baseUrl%>/member/show_update/<%$row.username%>'>修改</a>
-					<a href='<%$baseUrl%>/member/confirm_delete/<%$row.username%>'>删除</a>
-				</td>
+				<%if $is_admin eq 0%>
+					<td>
+						<a href='<%$baseUrl%>/member/show_update/<%$row.username%>'>修改</a>
+						<a href='<%$baseUrl%>/member/confirm_delete/<%$row.username%>'>删除</a>
+					</td>
+				<%/if%>
 			</tr>
 
 		<%/foreach%>
